Widget upgraded + Image updated
Finally fixed widget I figured out that problem in the previous image was really big sized and widget had memory leaks because of that. So I changed the image and it works now!! Also redisigned UI of widget a little bit.
Another big change is change is about calendar. It was showing days -1 so I fixed it. And finally I edited algorithms for streak. If today is 0 submissions then we wanna know if yesterday was 0 submissions too if yes then return 0 streak is yesterday was not 0 its a possibility that by the  end of the day will be submissions and return submissions not counting today. But if today is some submission then simply show the streak.